requirements.txt
streamlit
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
flask
asgiref
python-dotenv
botbuilder-core
botbuilder-schema
botbuilder-dialogs
botbuilder-ai
flask-cors
quart
hypercorn
aiohttp
asyncio
pyjwt
requests


